% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djan.R
\docType{data}
\name{djan}
\alias{djan}
\title{Cross-sectional data for countries}
\format{
An object of class `"tibble"`
}
\usage{
data(djan)
}
\description{
Cross-sectional data for countries
}
\examples{

data(djan)
head(djan)
}
\references{
https://link.springer.com/article/10.1007%2Fs11368-020-02633-z
}
\keyword{datasets}
