% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djan17.R
\docType{data}
\name{djan17}
\alias{djan17}
\title{Cross-sectional data for countries}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{ccode}{ISO3 country code (as character) for the countries in the
     sample (Angola, Benin, France, Rwanda, and UK)}
 \item{year}{A value between 1990 and 1999}
 \item{month}{An abbreviation (MMM) for month (character)}
 \item{gpd}{A fake value for GDP (randomly generated)}
 \item{population}{A fake value for population (randomly generated)}
}
}
\usage{
data(djan17)
}
\description{
Cross-sectional data for countries
}
\examples{

data(djan17)
head(djan17)
}
\references{
https://link.springer.com/article/10.1007%2Fs11368-020-02633-z
}
\keyword{datasets}
