% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djanturb.R
\docType{data}
\name{djanturb}
\alias{djanturb}
\title{Djankuat river turbidity logger measurements}
\format{
An object of class `"tibble"`
}
\usage{
data(djanturb)
}
\description{
Turbidity measurements at Djankuat river (Northen
Caucasus) in June 2016. Data recorded by \emph{Analite NEP-495}
}
\examples{

data(djanturb)
head(djanturb)
}
\references{
Chalov SR, Tsyplenkov AS. 2020. Influence of macroturbulence on the dynamics of river water turbidity. Vestnik Moskovskogo universiteta. Seriya 5, Geografiya 0 (3): 34–46 (In Russ.)
}
\keyword{datasets}
