% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_events.R
\name{hydro_events}
\alias{hydro_events}
\title{Splitting to hydrological events}
\usage{
hydro_events(dataframe, q, datetime, window = 1)
}
\arguments{
\item{dataframe}{A data set object}

\item{q}{Water discharge variable (or water stage)}

\item{datetime}{Datetime variable (column in POSIXct format)}

\item{window}{Indicate time period to search for a local minimum (in hours)}
}
\value{
A data frame object with a hydrological events column \code{he}
added
}
\description{
The main aim of this function is to split your dataset
into hydrological events based on the local minimum method
}
\examples{
data(djan)
output_table <- hydro_events(dataframe = djan, q = discharge, datetime = time, window = 21)
}
