% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHI.R
\name{AHI}
\alias{AHI}
\title{Calculate Aich's Hysteresis Index}
\usage{
AHI(dataframe, q, ssc)
}
\arguments{
\item{dataframe}{A data set object}

\item{q}{Water discharge variable}

\item{ssc}{Suspended sediment concentration variable}
}
\value{
a numeric value either NA
}
\description{
This function calculates Hysteresis Index proposed by \emph{Aich et al.} (2014)
}
\examples{
library(dplyr)
data(djan)
output_table <- hydro_events(dataframe = djan,
                             q = discharge,
                             datetime = time,
                             window = 21)

output_table \%>\%
  filter(he == 2) \%>\%
  AHI(q = discharge, ssc = SS)
}
