% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HImid.R
\name{HImid}
\alias{HImid}
\title{Calculate Lawler's Mid Hysteresis Index}
\usage{
HImid(dataframe, q, ssc, .warn = TRUE)
}
\arguments{
\item{dataframe}{data.frame object, containing water discharge
(`q`) and suspended sediment concentration (`ssc`) columns}

\item{q}{numeric, water discharge variable.}

\item{ssc}{numeric, suspended sediment concentration variable.}

\item{.warn}{logical, indicating if the warning message
should be displayed.}
}
\value{
a numeric value either NA
}
\description{
This function calculates Hysteresis Index
proposed by \emph{Lawler et al.} (2006)
}
\examples{
library(dplyr)
data(djan)
output_table <- hydro_events(dataframe = djan,
                             q = discharge,
                             datetime = time,
                             window = 21)

output_table \%>\%
  filter(he == 3) \%>\%
  HImid(q = discharge, ssc = SS)
}
\references{
Lawler DM, Petts GE, Foster IDL, Harper S. 2006.
Turbidity dynamics during spring storm events in an urban
headwater river system: The Upper Tame, West Midlands, UK.
Science of The Total Environment 360 (1): 109–126
DOI: 10.1016/j.scitotenv.2005.08.032
}
